requirements.txt
pytorch
matplotlib
pyvista